:recycle: [Docs]: Refactoring of Record Component
[비즈니스 로직 함수 분리]
- RecordOrder.tsx
- RecordOrderList.tsx
[ any ]
- button color change in RecordsRepairPopup
